move binarize_mask to ImageUtils
