+++ src/ImageUtils/ImageUtils.jl
@@ -15,7 +15,8 @@ export masker,
     imcomplement,
     to_uint8,
     get_tile_meta,
-    bump_tile
+    bump_tile,
+    binarize_mask
 
 include("brighten.jl")
 include("imcomplement.jl")
